Got rid of db info strings in each of the classes, now only using a boost property tree to move that info around. Also added a deleteConnection function to basicDbOps library so connections to the db wouldn't add up.

+++ Cpp/CXR_classify/storeupdater.cpp
@@ -24,7 +24,7 @@ void StoreUpdater::updateProgressBar()
     }
 
     emit attemptUpdateProBarValue(0);
-    while (bdo::countRecords(dbInfo, metadataTableName) < expected_num_files) {
+    while (((quint64) bdo::countRecords(dbInfo, metadataTableName)) < expected_num_files) {
         emit attemptUpdateProBarValue(bdo::countRecords(dbInfo, metadataTableName));
     }
     emit attemptUpdateProBarValue(bdo::countRecords(dbInfo, metadataTableName));
